Divide test suite in several files
Rebuild times were very affected by the 2.5K lines of the file.
Now is 8x faster.
